Changed GP Screen Space to operate in world space
Grease Pencil Screen Space used to operate based on the viewport,
which made previewing strokes very difficult and cumbersome.

This method gives the same end result, but works by scaling the
line thickness in world space with a scale factor proportional
to distance from the camera, resulting in lines that exist
in world space, but are rendered in screen space.

Overall, makes working with these lines a lot easier.

Modified screen space to operate by